requirements/base.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-autoslug==1.9.9
django-cities-light==3.9.2
django-jazzmin==2.6.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
phonenumbers==8.13.9
Pillow==9.5.0
progressbar2==4.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-utils==3.5.2
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
sqlparse==0.4.3
tzdata==2023.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.15
############################
django-recaptcha==3.0.0
django-filter==23.1
django-elasticsearch-dsl==7.3.0
